Painting Details class with other layouts
Added fav and cart menu items in the menu items folder.
The painting detail class shows the price of each painting with different sizes and multiple images can be seen loaded from the server. As well as reviews about the paintings can be checked in realtime and below the reviews, relative paintings are shown.